requirements.txt
flask
pymongo
python-dotenv
pandas
matplotlib
seaborn
jupyter
